+++ packages/_private/vanilla-demo/src/main.ts
@@ -23,13 +23,13 @@ import '@synergy-design-system/tokens/themes/sick2018_light.css';
 import '@synergy-design-system/components/index.css';
 import '@synergy-design-system/styles';
 import './app.css';
+import {
+  type AllowedSizes,
+  setGlobalSize,
+} from '@synergy-design-system/demo-utilities';
 
 import { createLayout } from './layout.js';
 import { initRouting } from './routing.js';
-import {
-  type AvailableSizes,
-  setGlobalSize,
-} from './utils.js';
 
 const initLayoutSwitch = async () => {
   await customElements.whenDefined('syn-button-group');
@@ -102,7 +102,7 @@ const initSizeSwitch = async () => {
       const clickedButton = target.closest('syn-icon-button') as SynIconButton;
       const size = clickedButton.dataset.size!;
 
-      setGlobalSize(size as AvailableSizes);
+      setGlobalSize(size as AllowedSizes);
 
       buttons.forEach(btn => {
         btn.color = btn.dataset.size === size ? 'primary' : 'currentColor';